@@ -7,33 +7,50 @@
   "repository": "jonathantneal/postcss-dir-pseudo-class",
   "homepage": "https://github.com/jonathantneal/postcss-dir-pseudo-class#readme",
   "bugs": "https://github.com/jonathantneal/postcss-dir-pseudo-class/issues",
-  "main": "index.js",
+  "main": "index.cjs.js",
+  "module": "index.es.js",
   "files": [
-    "index.js"
+    "index.cjs.js",
+    "index.es.js"
   ],
   "scripts": {
-    "clean": "git clean -X -d -f",
-    "prepublish": "npm test",
-    "test": "echo 'Running tests...'; npm run test:js && npm run test:tape",
-    "test:js": "eslint *.js --cache --ignore-pattern .gitignore",
+    "prepublishOnly": "npm test",
+    "pretest": "rollup -c .rollup.js --silent",
+    "test": "echo 'Running tests...'; npm run test:ec && npm run test:js && npm run test:tape",
+    "test:ec": "echint --ignore index.*.js test",
+    "test:js": "eslint *.js --cache --ignore-path .gitignore --quiet",
     "test:tape": "postcss-tape"
   },
   "engines": {
     "node": ">=4.0.0"
   },
   "dependencies": {
-    "browserslist": "^2.4.0",
-    "postcss": "^6.0.11",
-    "postcss-selector-parser": "^2.2.3"
+    "postcss": "^6.0.20",
+    "postcss-selector-parser": "^3.1.1"
   },
   "devDependencies": {
-    "eslint": "^4.7.1",
-    "eslint-config-dev": "2.0.0",
-    "postcss-tape": "2.1.0",
-    "pre-commit": "^1.2.2"
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.2",
+    "babel-preset-env": "^1.6.1",
+    "echint": "^4.0.1",
+    "eslint": "^4.19.0",
+    "eslint-config-dev": "^2.0.0",
+    "postcss-tape": "^2.2.0",
+    "pre-commit": "^1.2.2",
+    "rollup": "^0.57.1",
+    "rollup-plugin-babel": "^3.0.3"
   },
   "eslintConfig": {
-    "extends": "dev"
+    "extends": "dev",
+    "parser": "babel-eslint",
+    "rules": {
+      "complexity": [
+        "error",
+        {
+          "max": 30
+        }
+      ]
+    }
   },
   "keywords": [
     "postcss",
